+++ src/popup/pages/Send/tokenSign/index.tsx
@@ -2,7 +2,7 @@ import { DAPP_ACTION_CANCEL_ALL } from "@/constant/msgTypes";
 import { TOKEN_BUILD } from "@/constant/tokenMsgTypes";
 import useFetchAccountData from "@/hooks/useUpdateAccount";
 import Loading from "@/popup/component/Loading";
-import ICON_Arrow from "@/popup/component/SVG/ICON_Arrow";
+import SvgIcon from "@/popup/component/SvgIcon";
 import { updateShouldRequest } from "@/reducers/accountReducer";
 import {
   refreshTokenSignPopup,
@@ -245,14 +245,14 @@ const TokenSignPage = () => {
               $disabled={leftArrowStatus}
               onClick={onClickLeftBtn}
             >
-              <ICON_Arrow stroke={leftArrowColor} />
+              <SvgIcon src="/img/icon_chevron_left.svg" color={leftArrowColor} />
             </StyledMultiRowArrow>
             <StyledMultiRowArrow
               $disabled={rightArrowStatus}
               $isRight
               onClick={onClickRightBtn}
             >
-              <ICON_Arrow stroke={rightArrowColor} />
+              <SvgIcon src="/img/icon_chevron_left.svg" color={rightArrowColor} />
             </StyledMultiRowArrow>
           </StyledMultiTitleRowRight>
         </StyledMultiTitleRow>